source/requirements2.txt
astroid==1.6.1
cairocffi==1.1.0
CairoSVG==2.4.2
certifi==2018.10.15
cffi==1.14.0
chardet==3.0.4
#colorama==0.3.9 done
#colorlog==4.1.0 done
#colormap==1.0.3 done
cssselect2==0.3.0
defusedxml==0.6.0
Django==2.2.2
django-appconf==1.0.2
django-bootstrap-pagination==1.7.0
django-ckeditor==5.6.1
django-compressor==2.2
django-formtools==2.1
django-js-asset==1.1.0
django-libsass==0.7
django-localflavor==2.0
django-pagination-bootstrap==1.3
django-reversion==2.0.12
#easydev==0.9.38 done
html5lib==1.0.1
hurry.filesize==0.9
idna==2.7
isort==4.3.0
#Jinja2==2.11.1 done
lazy-object-proxy==1.3.1
libsass==0.13.5
logilab-common==1.4.1
#MarkupSafe==1.1.1 done
mccabe==0.6.1
npm==0.1.1
#numpy==1.18.2 done
oauthlib==2.1.0
optional-django==0.1.0
#pandas==1.0.3 done
#pexpect==4.8.0 done
Pillow==6.0.0
psycopg2==2.7.3.2
# ptyprocess==0.6.0 done
pycparser==2.20
pylint==1.8.2
Pyphen==0.9.5
#python-dateutil==2.8.1 done
pytz==2017.3
rcssmin==1.0.6
requests==2.20.1
requests-oauthlib==1.0.0
rjsmin==1.0.12
six==1.11.0
sqlparse==0.3.0
tinycss2==1.0.2
typed-ast==1.4.1
urllib3==1.24.3
WeasyPrint==51
webencodings==0.5.1
wrapt==1.11.2
